+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
-
+## [0.2.0]
+- Use `usbipd state` instead of `usbipd wsl list` to get the list of device
+- When the usb device is physically disconnected, pressing detached will hide the button if no more device is attached
 ## [0.1.0]
 
 - Initial release